バックエンドのユーザAPIを改修した
